+++ R/tar_config_set.R
@@ -97,7 +97,13 @@
 #' @param seconds_reporter Argument of [tar_make()], [tar_make_clustermq()],
 #'   and [tar_make_future()]. Positive numeric of length 1 with the minimum
 #'   number of seconds between times when the reporter prints progress
-#'   messages to the R console.
+#'   messages to the R console (for the aforementioned
+#'   [tar_make()]-like functions only).
+#' @param seconds_reporter_outdated Argument of [tar_outdated()]
+#'   and other related functions that do not run the pipeline.
+#'   Positive numeric of length 1 with the minimum
+#'   number of seconds between times when the reporter prints progress
+#'   messages to the R console for [tar_outdated()].
 #' @param shortcut logical of length 1, default `shortcut` argument
 #'   to [tar_make()] and related functions.
 #'   If the argument `NULL`, the setting is not modified.
@@ -177,6 +183,7 @@ tar_config_set <- function(
   seconds_meta_append = NULL,
   seconds_meta_upload = NULL,
   seconds_reporter = NULL,
+  seconds_reporter_outdated = NULL,
   seconds_interval = NULL,
   store = NULL,
   shortcut = NULL,
@@ -203,6 +210,7 @@ tar_config_set <- function(
   tar_config_assert_seconds_meta_append(seconds_meta_append)
   tar_config_assert_seconds_meta_upload(seconds_meta_upload)
   tar_config_assert_seconds_reporter(seconds_reporter)
+  tar_config_assert_seconds_reporter_outdated(seconds_reporter_outdated)
   tar_config_assert_seconds_interval(seconds_interval)
   tar_config_assert_shortcut(shortcut)
   tar_config_assert_store(store)
@@ -226,8 +234,10 @@ tar_config_set <- function(
     yaml[[project]]$seconds_meta_append
   yaml[[project]]$seconds_meta_upload <- seconds_meta_upload %|||%
     yaml[[project]]$seconds_meta_upload
-  yaml[[project]]$seconds_reporter <- seconds_reporter %|||%
-    yaml[[project]]$seconds_reporter
+  yaml[[project]][["seconds_reporter"]] <- seconds_reporter %|||%
+    yaml[[project]][["seconds_reporter"]]
+  yaml[[project]]$seconds_reporter_outdated <- seconds_reporter_outdated %|||%
+    yaml[[project]]$seconds_reporter_outdated
   yaml[[project]]$seconds_interval <- seconds_interval %|||%
     yaml[[project]]$seconds_interval
   yaml[[project]]$shortcut <- shortcut %|||% yaml[[project]]$shortcut
@@ -372,6 +382,9 @@ tar_config_assert_seconds_reporter <- function(seconds_reporter) {
   tar_assert_ge(seconds_reporter, 0)
 }
 
+tar_config_assert_seconds_reporter_outdated <-
+  tar_config_assert_seconds_reporter
+
 tar_config_assert_shortcut <- function(shortcut) {
   if (is.null(shortcut)) {
     return()
@@ -418,7 +431,7 @@ tar_reporters_make <- function() {
 }
 
 tar_reporters_outdated <- function() {
-  c("forecast", "silent")
+  c("forecast_interactive", "forecast", "silent")
 }
 
 tar_config_read_yaml <- function(config) {