+++ R/tar_network.R
@@ -82,7 +82,7 @@ tar_network <- function(
   exclude = NULL,
   outdated = TRUE,
   reporter = targets::tar_config_get("reporter_outdated"),
-  seconds_reporter = targets::tar_config_get("seconds_reporter"),
+  seconds_reporter = targets::tar_config_get("seconds_reporter_outdated"),
   callr_function = callr::r,
   callr_arguments = targets::tar_callr_args_default(callr_function, reporter),
   envir = parent.frame(),
@@ -92,6 +92,7 @@ tar_network <- function(
   force(envir)
   tar_assert_lgl(targets_only)
   tar_config_assert_reporter_outdated(reporter)
+  reporter <- tar_outdated_reporter(reporter)
   tar_assert_callr_function(callr_function)
   tar_assert_list(callr_arguments)
   tar_assert_dbl(seconds_reporter)